Fix Makefile
The `gen` bindings weren't doing anything.
